On branch test  Changes to be committed: 	new file:   .vscode/launch.json     new file:   .vscode/settings.json 	new file:   Resources/sshbanners.json 	modified:   honey.js 	new file:   improvedhoney.js 	new file:   test.js
Added improved honeyfile along with new features that will be fleshed out in test branch.
Working on adding ssh banner spoofing based on the server the script is ran on and a false
login prompt that leads the attacker on and gets as much data on them as possible before terminating
the connection and logging results.